feat: refactor movies page #9
