+++ src/pages/movies.astro
@@ -1,8 +1,28 @@
 ---
+import { signal } from "@preact/signals";
 import { MoviesList } from "../components/MoviesList";
 import MainLayout from "../layouts/MainLayout.astro"
+import { MovieType } from "../utils/types";
+import { getDBMovies, getMoviesData } from "../utils/getMovies";
+import { LoadingComponent } from "../components/ui/LoadingComponent";
+
+const movies = signal<MovieType[] | undefined>(undefined)
+const error = signal<string | undefined>(undefined)
+
+try {
+  const dbMovies = await getDBMovies()
+  const moviesIds = dbMovies.map(({ movieId }) => movieId)
+  movies.value = await getMoviesData(moviesIds)
+} catch(errorMessage) {
+  console.warn(errorMessage)
+  // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+  error.value = `${errorMessage}`
+}
+
 ---
 
 <MainLayout title='Movies'>
-  <MoviesList client:visible />
+  <LoadingComponent isLoading={!movies.value} error={error.value}>
+    <MoviesList movies={movies.value} />
+  </LoadingComponent>
 </MainLayout>
